Port `HDBSCAN` to `InteropMixin`/`ProxyBase` (#6916)
This ports our `HDBSCAN` to be based on `InteropMixin` instead of `UniversalBase`. It adds a test for sklearn interop to ensure we're converting correctly. It also moves the `cuml.accel` wrapper to be based on `ProxyBase` instead of `ProxyMixin`. Since this was the last model using the old `ProxyMixin` base class, we also delete `ProxyMixin` and the respective tests.

Fixes #6711.

Authors:
  - Jim Crist-Harif (https://github.com/jcrist)

Approvers:
  - Tim Head (https://github.com/betatim)

URL: https://github.com/rapidsai/cuml/pull/6916