Remove `Base.__getattr__`/`Base.__getstate__`/`Base.__setstate__`
`Base.__getattr__` existed to forward any missing attributes on a model
to the underlying `solver_model` attribute (if any). Beyond being
implicit, this led to nicer user error messages raised in properties on
subclasses being swallowed up as a generic `AttributeError`.

This PR removes the magic `__getattr__` in favor of explicitly surfacing
attributes in the parent model. The only place this was needed was the
`MBSGD*` estimators.

The `__getstate__`/`__setstate__` implementations on `Base` were only
needed due to the `__getattr__` implementation, since without them
`model.__getstate__` would return `model.solver_model.__getstate__`.
Removing these as well in favor of relying on python's standard behavior
(which matched the previous implementations).