Deprecate hdbscan prediction functions in `cuml.cluster` namespace
This finishes up a TODO in the codebase to remove these from the
`cuml.cluster` namespace. We just did a similar deprecation removing the
`cuml.cluster.hdbscan.prediction` namespace. The canonical import path
for all of these is now `cuml.cluster.hdbscan`.

Also fixes a typo (and adds a test) in the recent deprecation of the
`cuml.cluster.hdbscan.prediction` namespace.