Reflect output type in `hdbscan` prediction functions (#6939)
This implements type reflection for the output types of the `hdbscan` prediction functions.

For `all_points_membership_vectors`, the output type is the global output type (if set), falling back to the model output type. Since there is no input to this method, relying on the model output type makes the most sense to me.

For `membership_vector` and `approximate_predict`, the output type is the global output type (if set), falling back to the model output type, or the type of `points_to_predict` if `model.output_type == "input"` (the default). I _think_ this makes sense, since the freestanding functions take a model as their first argument they're effectively the same as methods on the model and should infer their output type the same IMO.

This is effectively a breaking change, since the output types of these functions will change. I don't see a way to deprecate this cleanly, and think the improvement is worth the API churn. Users passing in numpy arrays to any of these methods won't see a change, only those using other input types.

Fixes #6914.

Authors:
  - Jim Crist-Harif (https://github.com/jcrist)

Approvers:
  - Simon Adorf (https://github.com/csadorf)

URL: https://github.com/rapidsai/cuml/pull/6939