Compute `components_` in `DBSCAN`
This adds support for computing the `components_` attribute of `DBSCAN`.

The attribute is only stored if `calc_core_sample_indices=True` (the
default), since both are related to the core samples. The number of core
samples is much smaller than the number of samples (sometimes it's 0),
in practice this adds negligible time and space overhead, but will
respect the existing switch if needed.