Support invoking `cuml_accel_tests` without extra pytest flags (#6993)
This removes the need for adding a `-p cuml.accel` when invoking the `cuml_accel_tests` tests. This makes it easier for devs to run the tests just as they would other pytests.

The plugin is still needed to run the upstream test suites, but that's really what it's for anyway.

Also adds a small config option to ignore tests in the `cuml_accel_tests/upstream` directory. This won't affect CI (since there are no tests in that dir on a fresh clone), but improves the local dev experience since without it the upstream umap test suite would accidentally be collected when running the `cuml_accel_tests`.

Authors:
  - Jim Crist-Harif (https://github.com/jcrist)

Approvers:
  - Jake Awe (https://github.com/AyodeAwe)
  - Tim Head (https://github.com/betatim)

URL: https://github.com/rapidsai/cuml/pull/6993