Reduce instances of `_batched_kalman_device_loop_large_kernel` from 6 to 1
These instances are heavily templated based on a custom (only used in
ARIMA) linear algebra library. Mucking with this is a lot of code
movement.

The template parameters only dictate who the operations are batched and
parallelized though (meaning they're for performance, not correctness).
Since most ARIMA models should be small in parameters (i.e. the large
kernel loop shouldn't be hit much anyway), we choose a fixed
parameterization and move on. If ARIMA ever becomes important we can
revisit this, but right now we do the cheap thing.