Use `cuvs::neighbors::knn_merge_parts`
This moves to use `cuvs::neighbors::knn_merge_parts` instead of
`raft::spatial::knn::knn_merge_parts`. The latter was the 2nd largest
kernel used in cuml, and will also be going away in favor of an
equivalent version in `cuvs` (which we also link to). No sense in having
both in the binary.