Use `cuvs::neighbors::knn_merge_parts` (#7005)
This moves to use `cuvs::neighbors::knn_merge_parts` instead of `raft::spatial::knn::knn_merge_parts`. The latter was the 2nd largest kernel used in cuml, and will also be going away in favor of an equivalent version in `cuvs` (which we also link to). No sense in having both in the binary.

This drops 7.6 MiB in kernels from the `cuml` binary (for one SM).

Fixes #7002.

Authors:
  - Jim Crist-Harif (https://github.com/jcrist)

Approvers:
  - Victor Lafargue (https://github.com/viclafargue)

URL: https://github.com/rapidsai/cuml/pull/7005