Mark `Ridge` as not supporting underdetermined systems (#7003)
This addresses the `Ridge` part of #6864

cuSolver doesnt seem to support underdetermined problems (`n_samples < n_features`) and crashes when you attempt to use it. The long term solution might be to add support to cuSolver.

For now we raise `UnsupportedOnGPU`. This means we can remove a lot of xfails ~~and gain three. The three are related to handling `list` inputs. In general we don't handle that, I think, so I made no attempt here to deal with it either. WDYT?~~

Authors:
  - Tim Head (https://github.com/betatim)

Approvers:
  - Jim Crist-Harif (https://github.com/jcrist)
  - Simon Adorf (https://github.com/csadorf)

URL: https://github.com/rapidsai/cuml/pull/7003