Update title of `cuml.accel` documentation to "Zero Code Change Acceleration" (#7030)
I think that `cuml.accel` as a section header doesn't work well because it's not immediately clear what it means before learning about it. A more descriptive title would be better. From here on, we can refer to it more briefly as `cuml.accel` within the text.

I am also not a fan of using "code" as a section title, since `cuml.accel` is primarily a namespace and program name in the form of a CLI entrypoint.

Authors:
  - Simon Adorf (https://github.com/csadorf)

Approvers:
  - Jim Crist-Harif (https://github.com/jcrist)

URL: https://github.com/rapidsai/cuml/pull/7030