Add a CI run with "intermediate" versions of dependencies (#6988)
Towards https://github.com/rapidsai/cuml/issues/6677#issuecomment-3045842771
Closes https://github.com/rapidsai/cuml/issues/6999

We are claiming compatibility with a wide range of versions of scikit-learn, this adds a test for v1.5 to cover the middle between the two extremes of 1.4 and 1.7. The goal is to have a nightly (once a day) CI run that tests with an intermediate version of scikit-learn.

This is an attempt to explore options for doing this.

The shared workflows that we use do not have the notion of "intermediate" versions. The matrix only supports "oldest" and "latest". The simplest way to implement this would be to [add a new entry in that matrix](https://github.com/rapidsai/shared-workflows/blob/a8fe84c8249ec3553f271fe8ed3b5247543b3561/.github/workflows/conda-python-tests.yaml#L112). However, we would then need to find a way to filter it out for all repos that do not want this job. This seems quite complex/labour intensive.

In this PR we (ab)use `jq`'s ability to not only filter but rewrite things. We select the "oldest" deps entry and then rewrite it to say `"intermediate"`. Combined with an entry in `dependencies.yaml` that corresponds to that, we then get scikit-learn v1.5 more or less "for free". In particular we do not have to coordinate with repos outside of cuml about this customisation.

Authors:
  - Tim Head (https://github.com/betatim)

Approvers:
  - Simon Adorf (https://github.com/csadorf)
  - Gil Forsyth (https://github.com/gforsyth)
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/cuml/pull/6988