Rewrite `random_projection`
This fully rewrites the `cuml.random_projection` module to make use of
cupy instead of C++ from libcuml. The motivations for this were:

- The previous implementation didn't expose the fit `components_` array
  at the python level, meaning it was impossible to pickle the model.
  This made it impossible to save a fit model, something data scientists
  will want to do.
- The previous implementation also was written in a way that made it
  fail to work with `sklearn.clone`.
- The previous implementation didn't properly support sparse inputs,
  something the sklearn implementation did.

While all of these could have been fixed while keeping the
implementation in C++ (with some major refactoring), there also wasn't a
strong reason for the implementation to be done in C++ anyway. The
models here are simple and may be implemented straightforwardly using
cupy primitives without added performance cost. This reduces our
maintenance burden while also fixing bugs.

Note that this doesn't come with a performance cost. Rather, a quick
benchmark shows we're roughly 4x faster for fitting
`GaussianRandomProjection`, and 14x faster for fitting
`SparseRandomProjection`. For `transform` we're roughly the same to
slightly faster (especially for sparse inputs). All this with less code
and more features.