FIX Make GaussianNB more resilient (#7113)
This makes the minimal changes needed so that `GaussianNB` does not end up in a broken state where all common checks lead to a `MemoryError`. The basic fix was to check that `X` and `y` have the same number of samples and erroring out (instead of ploughing on).

The sad thing is that to find this fix you had to run all the checks, find the first `MemoryError` and then run that check with an increasing number of the checks that run before it. Once the memory error raises its head you know that something in the most recently added check must be causing a problem.

towards #7100

Authors:
  - Tim Head (https://github.com/betatim)

Approvers:
  - Jim Crist-Harif (https://github.com/jcrist)

URL: https://github.com/rapidsai/cuml/pull/7113