Nicer error message for not implemented fitted attributes
`cuml.accel` doesn't support every fitted attribute that sklearn does.
For many users this won't matter - many of the fitted attributes aren't
accessed for typical usage (some only exist to hold state and aren't
something a user would want to see; others are only used for
introspecting results in rare circumstances).

Currently we just raise the default `AttributeError` for these
attributes, which doesn't make it clear to the user what went wrong (was
it their code? Did they have a typo? Was it cuml.accel?).

With this PR we raise a nicer error message telling users that the
attribute isn't implemented, and asking them to open an issue if this
attribute is necessary for their use case.