Note non-determinism in random projection transforms
Due to an upstream cupy bug, mixing sparse/dense arrays may result in
close but not perfectly reproducible results. This means that passing a
sparse array to `GaussianRandomProjection.transform` (or a dense array
to `SparseRandomProjection.transform`) won't produce exactly identical
results each run, even with a `random_state` set.

For now we document this limitation and relax the bounds on a flaky
test.