Nicer error message for not implemented fitted attributes in `cuml.accel` (#7127)
`cuml.accel` doesn't support every fitted attribute that sklearn does. For many users this won't matter - many of the fitted attributes aren't accessed for typical usage (some only exist to hold state and aren't something a user would want to see; others are only used for introspecting results in rare circumstances).

Currently we just raise the default `AttributeError` for these attributes, which doesn't make it clear to the user what went wrong (was it their code? Did they have a typo? Was it cuml.accel?).

With this PR we raise a nicer error message telling users that the attribute isn't implemented, and asking them to open an issue if this attribute is necessary for their use case.

With this I _think_ the pressure for finishing up all missing attributes noted in #6966 decreases. We should still fill these in as best as we can, but having a nicer user-facing error message should help users understand what went wrong (and if they open issues, help us understand the importance of some of these that are missing).

Part of #6966.

Authors:
  - Jim Crist-Harif (https://github.com/jcrist)

Approvers:
  - Tim Head (https://github.com/betatim)
  - Simon Adorf (https://github.com/csadorf)

URL: https://github.com/rapidsai/cuml/pull/7127