Don't force `n_streams=1` if `random_state` is set
I've experimented with a bunch of different inputs/distributions and I
can't find a case of non-deterministic output even if `n_streams=4`.
There might be one, but we'd need someone more familiar with this code
to verify. For now it seems sufficiently safe to remove this guard in
`cuml.accel`, I can't find a case where it would matter.