`n_streams=1` is not required for determinism in `cuml.ensemble` (#7140)
I've experimented with a bunch of different inputs/distributions and I can't find a case of non-deterministic output even if `n_streams=4`. There might be one, but we'd need someone more familiar with this code to verify.

For now it seems sufficiently safe to remove forcing `n_streams=1` if `random_state is not None` in `cuml.accel`. Forcing `n_streams=1` is slower, and doesn't seem to matter for any workload I've tried.

This also removes the caveats in the docs for `cuml.ensemble` proper.

Fixes #7111.

Authors:
  - Jim Crist-Harif (https://github.com/jcrist)

Approvers:
  - Tim Head (https://github.com/betatim)

URL: https://github.com/rapidsai/cuml/pull/7140