Fix HTML representation's fitted status (#7146)
Need to sync the attributes to the CPU model for the HTML repr to work properly. This is because we delegate the HTML repr building to the CPU model. The advantage of this is that users get exactly what they are expecting, the downside is that we need to sync fitted attributes so that the fitted/not fitted detection works correctly. In theory we could dig into how the detection is done and only sync "that one attribute" needed to pass the test. I think that is a premature optimisation and relies on the internals of scikit-learn which means it is fragile.

closes #7145

Authors:
  - Tim Head (https://github.com/betatim)

Approvers:
  - Jim Crist-Harif (https://github.com/jcrist)

URL: https://github.com/rapidsai/cuml/pull/7146