Use custom plugin to download test data early (#7169)
~~Running this in CI as I can't reproduce the failure locally~~

The problem we were seeing in the naive bayes tests was that some test functions saw only a subset of the dataset. As a result there was not at least one sample from every class in the dataset. The reason this happened is some kind of race in the downloading and processing of the data. This happens because we use more than one worker for `pytest-xdist`.

We work around this problem by defining a custom plugin that runs before workers start and downloads all datasets.

The downside of this approach is that we need to manually list the datasets that get "pre-downloaded". I think that is Ok because we don't add new datasets frequently. But this could be improved.

An upside is that we only download each dataset once, not once per worker as we were doing so far.

More discussion and details in https://github.com/scikit-learn/scikit-learn/issues/32095 - maybe it is possible to fix this at the level of scikit-learn. Which would be great as it would mean we can remove this plugin again.

xref https://github.com/rapidsai/cuml/issues/7152

Authors:
  - Tim Head (https://github.com/betatim)

Approvers:
  - Simon Adorf (https://github.com/csadorf)
  - James Lamb (https://github.com/jameslamb)
  - Kyle Edwards (https://github.com/KyleFromNVIDIA)

URL: https://github.com/rapidsai/cuml/pull/7169