Move cert configuration to `pytest_configure` hook
Previously we had a `pytest_sessionstart` hook setup to install certifi
certs to be used by `urllib` when downloading datasets. Recently we also
added a `pytest_configure` hook to predownload the datasets at the start
of the test run.

`pytest_configure` runs before `pytest_sessionstart`, which means the
downloads were no longer using the certifi certs. This PR moves the cert
install to also be part of `pytest_configure` and explicitly earlier
than the downloads, fixing the problem.

Also removes the duplicate hook in the dask conftest file, there's no
reason for that to be there, the parent `conftest.py` is already used.