Remove `get_json`/`get_detailed_text`/`get_summary_text` from ensemble estimators (#7177)
This removes:

- `get_json`
- `get_detailed_text`
- `get_summary_text`

From the `cuml.ensemble` and `cuml.dask.ensemble` estimators. This is a breaking change, done without a deprecation period.

The primary use case for these methods was debugging and testing the internal state of the estimators. This should now be done through the use of treelite (`model.convert_to_treelite_model().dump_as_json()`). The schema of this output differs from that of the old `get_json`, but:

- it is a superset of the information available in the prior format
- it's standardized as part of treelite (rather than the old, undocumented internal representation we used)
- it can be used to reload the model (the previous `get_json` had no way to reconstruct a model from json)

Since we cannot _easily_ massage the output of the new treelite methods into the old undocumented schema (and we suspect usage of these methods in the wild is minimal-to-none) we drop them as a fully breaking change. Doing this lets us further cleanup the internals of our random forest estimators, fixing bugs and improving maintainability.

Fixes #7159.

Authors:
  - Jim Crist-Harif (https://github.com/jcrist)

Approvers:
  - Simon Adorf (https://github.com/csadorf)

URL: https://github.com/rapidsai/cuml/pull/7177