Small cleanup to `cuml.dask.metrics.confusion_matrix` (#7181)
We're seeing some odd flaky failures with this function. I'm unable to reproduce the issue locally, but it _might_ be due to serialization issues (I cannot explain why those would happen now, dask's serialization should be fine with how things were written before).

However, the way this function is written results in some unnecessary compute calls. Technically all of `confusion_matrix` could be done using stock, lazy, dask operations. However, since the current method computes eagerly and returns a cupy array that would be an API change. Instead, I swap out the part of the code that seems to be at least loosely tied to the errors we're seeing with more standard dask operations. Under the hood dask will use cupy if the input array is a device array, so this is functionally the same as what was there before, just using dask calls to do it. Not sure if this will fix the issue we're seeing in CI, but it's a valid change either way.

Possibly fixes #7179.

Authors:
  - Jim Crist-Harif (https://github.com/jcrist)

Approvers:
  - Simon Adorf (https://github.com/csadorf)

URL: https://github.com/rapidsai/cuml/pull/7181