Expose some private attrs in `cuml.accel`
Previously we never exposed private attributes in `cuml.accel`'s proxy
implementation. In most cases this is the correct choice, however some
estimators have private attributes that downstream consumers may still
use.

Previously we exposed those as a one-off way. We now add a builtin
mechanism, and expand support a bit. In particular, the `neighbors`
classifiers now expose the `_tree` attribute (it's always `None`, but
that should be supported by consumers).

This was motivated by getting `cuml.accel` to work with `PyOD`.