Accelerate `KernelDensity.score_samples`
This does away with the handwritten numba kernel entirely, replacing it
with a `cupy.ReductionKernel` (until `cp.logaddexp.reduce` is
implemented). This scales _much_ better than the old implementation
(5000x on large inputs when benchmarking just this part of the kernel).
For the whole `score_samples` routine we see a more modest 2-5x speedup
on typical problem sizes.