Use `rapids-xgboost` and exclude libnccl from auditwheel check (#7231)
Some of our builds were using an older version of xgboost that didn't work with the recent ABI breakage of `librmm`.

To fix this we move to using `rapids-xgboost` instead of `xgboost`. This pins `xgboost` to a specific rapids-supported-version for the current release cycle.

Additionally, we exclude `libnccl.so.*` from the auditwheel repair call since NCCL is a transitive dependency.

Fixes #7230.

Authors:
  - Jim Crist-Harif (https://github.com/jcrist)

Approvers:
  - James Lamb (https://github.com/jameslamb)
  - https://github.com/jakirkham

URL: https://github.com/rapidsai/cuml/pull/7231