`KernelDensity.score_samples` output type & dtype fixes (#7240)
Previously `KernelDensity.score_samples`:

- Always returned a `cupy` array, instead of following our `output_type` handling conventions.
- Returned inconsistent `dtype`, depending on the kernel specified.

We now ensure that `output_type` is followed, and that the output type matches that of the original `X` input to `KernelDensity.fit` (matching our dtype conventions elsewhere).

This also fixes the overflow in a test function, which was caused by returning a `float32` from a `float64` input, leading to overflows in the test. To further avoid this I've updated the test code to apply float64 anyway just to be sure. To be clear - the overflow here was in our test code, not in `KernelDensity` itself.

Fixes #7236.

Authors:
  - Jim Crist-Harif (https://github.com/jcrist)

Approvers:
  - Simon Adorf (https://github.com/csadorf)

URL: https://github.com/rapidsai/cuml/pull/7240