Fully reorganize `cuml.ensemble` estimator structure
This is a pretty clean rewrite of the internals. Before we'd always
store pointers to some internal state on `RandomForest*` as `int`s, even
though these pointers were really only necessary to eventually convert
the internal model structure into treelite and eventually into FIL for
inference. This handling of memory was both incorrect (leading to
memory leaks, segfaults) and inefficient (holding around internal GPU
state longer than it needed to be).

This also fixes a bug where the intermediate treelite handle created was
never freed, leading to a memory leak (potentially large, depending on
the size of the fit model).

We now fit and convert to treelite all in one go. This is simpler from
the caller side, and removes a bunch of memory footguns.

While at it, we also update the internals of all the `cuml.ensemble`
estimators to match current cuml and sklearn conventions. In particular,
these estimators no longer have nontrivial `__init__` methods, and don't
modify hyperparameters inplace. This greatly improves our sklearn
compatibility.