No need to serialize `treelite.Model` when loading FIL
Previously creating a new instance of `ForestInference` would
serialize/deserialize a model always, even if it was always loaded in
memory as a `treelite.Model`. There's no reason to do this, we can
pass in the loaded model directly already.

Further, we can rely on the python `treelite` bindings directly rather
than making native `treelite` calls. This helps simplify the memory
management and code.