Revert sharing treelite model handles
Storing a loaded `treelite.Model` handle created from `libcuml` and
wrapped in `treelite.Model` only works if `treelite` and `cuml` both are
using the same `libtreelite.so`. This is true for conda environments,
but not true for wheels where `libtreelite` and `libcuml` each have
their own version of this library. Sharing handles the way I had wrote
before was fine on conda, but caused problems for wheels (mostly when
releasing memory, since the library versions end up using different
allocators). For now we revert back to the less efficient and dance of
serialize-deserialize to convert between cuml's treelite and the python
library's treelite.