Ensure sklearn roundtrip attribute consistency
This was motivated by a segfault in our sklearn test suite that turns
out to be due to `SpectralEmbedding` having a C-contiguous array where
an F-contiguous array was expected.

While fixing this, I added a generic test for attribute consistency, and
fixed a few minor occurrences where we weren't 100% correct.