Ensure finite values passed to SpectralEmbedding (#7280)
The C++ implementation doesn't support non-finite values. Sometimes this results in a failure-to-converge error, sometimes this results in an illegal memory access (and downstream failures eventually leading to potential segfaults).

I _believe_ this to be the root cause of #7274 (if we see a failure, it's preceded by illegal memory accesses here). I've run things locally a bunch after this fix and haven't seen a failure, going to try it in CI.

Fixes #7274.

Authors:
  - Jim Crist-Harif (https://github.com/jcrist)

Approvers:
  - Simon Adorf (https://github.com/csadorf)

URL: https://github.com/rapidsai/cuml/pull/7280