Don't infer dtype from model in `cuml.explainer`
Previously the classes in `cuml.explainer` would try to infer the dtype
of the model function from `func.__self__.dtype`. This only worked for
some models (only those that exposed an undocumented `.dtype`
attribute), and led to inconsistent behavior between model functions. As
we cleanup our model internals we're also removing the `.dtype`
attribute from models, so this method no longer works.

Aiming for consistency, we change the default of `dtype` in
`cuml.explainer` to `float32` everywhere - if a user wants to use
`float64` they'll need to specify it manually.