Fix issue in Dask DBSCAN (#7359)
Closes #7341

Each worker processes the full dataset, and the resulting labels are merged toward the rank 0 through the communicator. The rank zero (only) then transforms the sentinel values into -1. However, the `internal_model` in charge of returning the final labels was selected as the first one in the alphabetical ordering of Dask workers IP addresses and not as the first rank in RAFT communicators clique. This caused the labels to be incorrect if the rank and alphabetic order did not coincide.

The PR also fixes an issue in the Cython code.

Authors:
  - Victor Lafargue (https://github.com/viclafargue)

Approvers:
  - Jim Crist-Harif (https://github.com/jcrist)

URL: https://github.com/rapidsai/cuml/pull/7359