Fix overflow in coordinate descent (#7399)
Previously our coordinate descent solver would fail on problems where `n_cols * n_rows > INT_MAX` due to an `int` overflow.

There were two locations where this was happening:

- Calculating the offset into the input matrix
- Within the computation of the L2Norm

The former is a quick local fix. The latter I also fixed locally by switching from an `int` to a `int64_t` in the template call. However, I'm not sure if that's the best fix, or if it'd be better to handle this upstream within the template itself to avoid overflow of the index types. This was easy to do so I did it here for now.

I've checked, and with this we can solve very large coordinate descent problems, with the dimension limitation now being `INT_MAX` in both rows and columns. Moving larger than that would require using the 64 bit cublas API, but I have no need for that now.

Additionally, on the python side if a user tries to pass a larger value they'll get a nicer python-side error, rather than a cublas error code (and a potentially corrupted handle).

Fixes #6736.

Authors:
  - Jim Crist-Harif (https://github.com/jcrist)

Approvers:
  - Victor Lafargue (https://github.com/viclafargue)

URL: https://github.com/rapidsai/cuml/pull/7399