Fix Dask DBSCAN tree reduction (#7398)
Closes #7341

The tree reduction operation of Dask DBSCAN had an issue. The communicator was waiting for communication request completion even when none was ordered (passive rank) causing a deadlock.
Additionally, there was an issue with the UCX comms being set with an incorrect number of ranks when Dask workers fail to register to the scheduler.

Authors:
  - Victor Lafargue (https://github.com/viclafargue)

Approvers:
  - Jim Crist-Harif (https://github.com/jcrist)
  - Divye Gala (https://github.com/divyegala)

URL: https://github.com/rapidsai/cuml/pull/7398