Deprecate `normalize` option
This deprecates the `normalize` option to all linear models that still
supported it.

It includes one _potentially_ breaking change.
`cuml.experimental.linear_model.Lars` was documented to have `normalize`
default to `False` (and clearly should have), but still had it default
to `True`. Since it's in the experimental namespace and feels like a bug
to me I've updated the value here to `False` with no deprecation period.
The tests had to have their tolerance slightly expanded.

AFAICT no model was testing (intentionally) with `normalize=True`
anywhere, which makes me feel pretty confident that we should rip this
option out. There's no way to ensure it was even doing the right thing
anymore.