Bring back `KMeansMG`
During a recent refactor we removed the `KMeansMG` class, viewing it as
internal. It turns out this class was used by a few external projects.

Since we still need to support external users accessing the non-dask
multi-gpu implementation, we'll want a public way to do so that isn't
the private `_fit` method. Additionally, since we want to special case
the `MG` case a little more, making it a separate class (even if as a
thin shim) makes sense.

This PR:

- Brings back the `KMeansMG` class
- Adds a check that `random_state` is non-None in the `KMeansMG` case,
  ensuring external users also set `random_state` properly
- Removes mutation of kwargs in the dask `KMeans` case
- Simplifies and moves the multi-gpu `kmeans++`/`oversampling_factor`
  check.