Several improvements to `Ridge` (#7410)
This PR includes several improvements to `Ridge`. I started trying to fix one issue, but they were all so coupled it's a bit of a large PR. Hopefully still understandable though.

- Move SVD solver to pure python. This simplifies maintenance, and improves flexibility. The new python-based solver is as fast or faster than the old C++ version, which makes sense since it's (mostly) just a thin layer over cublas and cusolver calls.
- Adds support for multi-target regression to `Ridge`. Fixes #4412.
- Adds support for `X` with more features than samples to `Ridge` with `solver="svd"`. Fixes #7198.
- Adds support for array-like `alpha`, improving our sklearn compatibility. Fixes a long standing TODO in our docs.
- Adds a `copy_X` parameter, mirroring the sklearn equivalent. This lets the solver mutate X instead of making a copy, reducing memory usage.
- Fixes a bug where the solver would accidentally mutate `y` and `sample_weight` in some configurations, and adds a test.
- Moves `Ridge` tests out into their own file, the `test_linear_models.py` file was getting a bit untenable.
- Vastly improved testing coverage and hygiene
- Cleanup docstrings to match conventions

It does include one breaking change, which I think we want to do based on other conversations. Previously if a user explicitly passed `solver="eig"` but the `"eig"` solver couldn't handle the inputs, it'd warn and fallback to `"svd"`. We now error in this case, alerting the user that the solver they specified isn't supported. This better matches sklearn conventions, and is something I think we'll want to apply to the `LinearRegression` implementation as well (see https://github.com/rapidsai/cuml/pull/7355#discussion_r2439081600). With the default of `solver="auto"` we'll still fallback to `"svd"` in those cases, we only fail if the user _explicitly requested_ a solver that doesn't support the input types.

Fixes #4412. Fixes #7198. Followup to #7330.

Authors:
  - Jim Crist-Harif (https://github.com/jcrist)

Approvers:
  - Simon Adorf (https://github.com/csadorf)

URL: https://github.com/rapidsai/cuml/pull/7410