Expose `n_iter_` in SVC/SVR
This PR exposes `n_iter_` in our `SVC` and `SVR` estimators.

While doing so, we also deprecate the old meaning of `max_iter` (max
outer iterations) in favor of a new meaning (max _total_ iterations).
This new meaning aligns with sklearn's implementation, easing interop
between the two.

The deprecation cycle is implemented as:

- 25.12: If a user passes in a non-default integer value, a warning
  is raised informing them of the deprecated meaning. The old behavior
  is still applied. To opt in to the new behavior (and silence the
  warning) they can pass an instance of `TotalIters` (an `int` subclass)
  instead. The `n_iter_` attribute reflects the number of total
  iterations.
- 26.02: The `TotalIters` wrapper will be deprecated. A bare integer
  will now mean "max total iterations".
- 26.04: The `TotalIters` wrapper will be fully removed.