Expose `penalty_normalized` and `lbfgs_memory` in `LogisticRegression`
Adds two new parameters to `LogisticRegression`/`LogisticRegressionMG`:

- `penalty_normalized`
- `lbfgs_memory`

There's no change in default behavior; the defaults remain the same as
before, this just exposes the parameters to forward to the QN solver to
the user. These parameters should rarely need to be used, but were
requested by the `spark-rapids-ml` team to better match Spark-MLLib
behavior.