Remove `target_dtype` and supporting infra (#7516)
With all the work in #7317, we're now at a point where `target_dtype` is no longer used.

This PR removes `target_dtype` and all supporting infra. This simplifies our decorators and base class, and reduces the amount of state stored on an estimator. Since this was all private implementation details, this is not a breaking change.

Authors:
  - Jim Crist-Harif (https://github.com/jcrist)

Approvers:
  - Anupam (https://github.com/aamijar)
  - Simon Adorf (https://github.com/csadorf)

URL: https://github.com/rapidsai/cuml/pull/7516