Remove verbose descriptor (#7527)
Previously the `verbose` attribute on an estimator had some magic in it where the value when accessed internally would be a `cuml.internals.logger.level_enum`, but externally would be an `int` or `bool` value as described by the public API.

This PR removes that magic in favor of simpler mechanisms. Now `model.verbose` is a normal attribute, and `model._verbose_level` a property that converts `model.verbose` to a `level_enum`.

Part of #5022.

Authors:
  - Jim Crist-Harif (https://github.com/jcrist)

Approvers:
  - Simon Adorf (https://github.com/csadorf)

URL: https://github.com/rapidsai/cuml/pull/7527