Rewrite GaussianNB to not use custom kernels
This is work by Cursor instructed to look at the cuml and scikit-learn
implementations and modify the cuml one based on what is done in
scikit-learn. I've done some manual review of the work, but mostly
looked at performance and interested to see if this solves the memory
errors we are seeing.