Deprecate `TotalIters` for SVC/SVR
This is the 2nd part of the deprecation plan for changing the behavior
of `max_iter` for `SVC`/`SVR`. Previously we added a `TotalIters`
wrapper so users could opt-in to the new behavior, now we make the new
behavior the default and deprecate the wrapper. `max_iter` now always
places a limit on _total iterations_ in the solver.